server/requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
jinja2
python-multipart
pyuvc @ git+https://github.com/pupil-labs/pyuvc.git
pytest

